Set thresholds at which review checks should fail

This function is experimental and as the key package metrics evolve, parameters may be subject to change.

set_thresholds(
  exports_without_examples = "fail",
  fk_fail = 30,
  fk_warn = 50,
  length_fail = 3000,
  length_warn = 2000
)

Arguments

exports_without_examples
Action to take when exports without examples are identified. Possible values are "fail", "warn", or "none".

fk_fail
Vignette Flesch-Kincaid scores lower than this will result in a review check failure

fk_warn
Vignette Flesch-Kincaid scores lower than this but higher than fk_fail will results in a review check warning

length_fail
Vignette word count longer than this will result in a review check failure

length_warn
Vignette word count longer than this but shorter than length_fail will result in a review check warning

Examples

set_thresholds(exports_without_examples = "warn")

#> $functions
#> $functions$exports_without_examples
#> [1] "warn"
#> 
#> 
#> $vignettes
#> $vignettes$fk
#> $vignettes$fk$fail
#> [1] 30
#> 
#> $vignettes$fk$warn
#> [1] 50
#> 
#> 
#> $vignettes$length
#> $vignettes$length$fail
#> [1] 3000
#> 
#> $vignettes$length$warn
#> [1] 2000
#> 
#> 
#>